+++ Source/MyGame/Private/Modules/Render/Prefabs/Sprite.h
@@ -20,7 +20,7 @@ struct SpriteParams
     sf::Vector2f position = {0.f, 0.f};
     sf::Vector2f scale = {1.f, 1.f};
     float rotation = 0.f;
-    int zOrder = 0;
+    float zOrder = 0;
 };
 
 struct Sprite
